pyproject.toml
[project]
name = "nttdatarag"
version = "0.1.0"
description = ""
authors = [
    {name = "anil",email = "ergananil@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pymupdf (>=1.26.3,<2.0.0)",
    "openai (>=1.97.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "jsonlines (>=4.0.0,<5.0.0)",
    "faiss-cpu (>=1.11.0.post1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "pdfminer-six (>=20250506,<20250507)",
    "pillow (>=11.3.0,<12.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
]

# pytest.ini
[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
